% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_combine_for_report.R
\name{combine_for_report}
\alias{combine_for_report}
\title{Combine multiple objects for a report}
\usage{
combine_for_report(...)
}
\arguments{
\item{...}{Objects to be combined.}
}
\value{
A combined object.
}
\description{
This function takes an arbitrary number of objects as arguments and combines them
into a single object. The combined object can then be used for further analysis
or reporting.
}
