% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_get_bibliography_measures.R
\name{get_bibliography_measures}
\alias{get_bibliography_measures}
\title{Used in the pipeline - Get Bibliography Measures}
\usage{
get_bibliography_measures(bibliography_fn)
}
\arguments{
\item{bibliography}{A data frame containing the bibliography dataset.}
}
\value{
A list containing the calculated measures.
}
\description{
This function calculates various measures related to a bibliography dataset.
}
