% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups}
\alias{groups}
\title{Zotero Groups to be Analysed}
\usage{
groups()
}
\description{
This function returns the Zotero groups and their IDs to be used in other functions.
}
\examples{
\dontrun{
groups()
}

}
