% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPBES.LiteratureReport-package.R
\docType{package}
\name{IPBES.LiteratureReport-package}
\alias{IPBES.LiteratureReport}
\alias{IPBES.LiteratureReport-package}
\title{IPBES.LiteratureReport: Literature Reports for IPBES Zotero Libraries}
\description{
Generate Reports for the literature in IPBES Zotero Libraries.
}
\author{
\strong{Maintainer}: Rainer Krug \email{Rainer@krugs.de} (\href{https://orcid.org/0000-0002-7490-0066}{ORCID})

}
\keyword{internal}
