% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPBES.LiteratureReport-package.R
\docType{package}
\name{LitReport-package}
\alias{LitReport}
\alias{LitReport-package}
\title{LitReport: Creation of literature reports as well as website and publishing it}
\description{
More about what it does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
