% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_load_bibliography.R
\name{load_bibliography}
\alias{load_bibliography}
\title{Used in the pipeline - Load and prepare bibliography}
\usage{
load_bibliography(
  bibliography_zotero = NULL,
  bibliography_name = "bibliography",
  bibliography_url = "",
  verbose = FALSE
)
}
\arguments{
\item{bibliography_zotero}{The file path to the bibliography exported from Zotero in CSV format.}

\item{bibliography_name}{The name of the bibliography.}

\item{bibliography_url}{The URL of the bibliography.}

\item{verbose}{Whether to display verbose output.}
}
\value{
A list containing the calculated metrics for the bibliography:
\itemize{
\item timestamp: The timestamp when the bibliography was created.
\item name: The name of the bibliography.
\item url: The URL of the bibliography.
\item bibliography: The data frame containing the bibliography.
\item dois_bib: The DOIs extracted from the bibliography.
\item isbns: The ISBNs extracted from the bibliography.
\item issns: The ISSNs extracted from the bibliography.
\item works: The works fetched using the DOIs.
\item dois_works: The standardized DOIs extracted from the works.
}
}
\description{
This function loads and prepares the bibliography for rurther analysis.
}
\examples{
\dontrun{
bibliography <- load_bibliography(
    bibliography_zotero = "path/to/bibliography.csv",
    bibliography_name = "My Bibliography",
    bibliography_url = "https://example.com/bibliography",
    verbose = TRUE
)

# Print the calculated metrics
print(bibliography)
}

}
