% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maks_all.R
\name{make_all}
\alias{make_all}
\title{Make all}
\usage{
make_all(overwrite = TRUE)
}
\arguments{
\item{overwrite}{If \verb{TRUE, the file }index.html\verb{will be overwritten. Default is}TRUE`.}
}
\value{
Return value from quarto::quarto_render
}
\description{
This function renders the document \code{index.qmd}. This includes:
\itemize{
\item install "devtools" and all dependencies needed for the code to run
\item dowlnoad the IPBES Zotero Libraries and store them as \code{.csv} files in ther folder \code{input}
\item generate the measures and graphs for the reports from all Zotero Libraries in the folde \verb{_targets}
\item generate the reports in the folder \code{output}
\item create or update the file \code{index.html} which contains all the links to the reports
}
}
\examples{
\dontrun{
make_all(overwrite = FALSE)
}
}
