% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_reports}
\alias{make_reports}
\title{Generate Reports for Assessments}
\usage{
make_reports(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Logical value indicating whether to overwrite existing reports. Default is \code{FALSE}.}
}
\description{
This function generates reports for assessments based on the provided bibliography files in the
folder \code{input}. The reports are generated using the \code{bibliography_report.qmd} Quarto document and
will bie located in the \code{output} folder.
}
\examples{
\dontrun{
# Generate reports without overwriting existing reports
make_reports()

# Generate reports and overwrite existing reports
make_reports(overwrite = TRUE)
}
}
