% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_make.R
\name{pipeline_make}
\alias{pipeline_make}
\title{Run a targets pipeline from the package}
\usage{
pipeline_make(dir = ".", ...)
}
\arguments{
\item{dir}{Path to the directory where the pipeline should run. Default is the current directory.}

\item{...}{Additional arguments passed to \code{targets::tar_make()}.}
}
\description{
Runs \verb{pipeline_prepare() and then runs }tar_make()` there.
}
