% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_prepare.R
\name{pipeline_prepare}
\alias{pipeline_prepare}
\title{Prepare folder to run targets pipeline from the package}
\usage{
pipeline_prepare(dir = ".", template_dir = NULL, overwrite = TRUE)
}
\arguments{
\item{dir}{Path to the directory where the pipeline should run. Default is the current directory.}

\item{overwrite}{Logical. Overwrite an existing _targets.R in \code{dir}? Default is FALSE.}

\item{script}{Optional path to a custom _targets.R file. If NULL, uses the package default.}

\item{groups_file}{Optional path to a custom groups.csv file. If NULL, uses the package default.}
}
\description{
Copies a default \verb{_targets.R} and \code{groups.csv} file to the specified directory (if not already present).
}
