% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_visnetwork.R
\name{pipeline_visnetwork}
\alias{pipeline_visnetwork}
\title{Plot a targets pipeline as a visNetwork graph}
\usage{
pipeline_visnetwork(dir = ".", ...)
}
\arguments{
\item{dir}{Directory containing the `_targets.R` file and pipeline metadata.}

\item{...}{Additional arguments passed to `targets::tar_visnetwork()`}
}
\value{
A `visNetwork` object (plotted interactively).
}
\description{
This function temporarily runs `tar_visnetwork()` in the specified directory
using the `_targets.R` file found there.
}
