% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_oa_status.R
\name{plot_oa_status}
\alias{plot_oa_status}
\title{Used in the pipeline - Plot Open Access Status by Publication Year}
\usage{
plot_oa_status(bibliography)
}
\arguments{
\item{bibliography}{A bibliography object containing information about publications.}
}
\value{
A ggplot object representing the stacked bar chart.
}
\description{
This function takes a bibliography object and plots the distribution of open access status
by publication year using a stacked bar chart.
}
\examples{
\dontrun{
bibliography <- read_bibliography("path/to/bibliography.csv")
plot_oa_status(bibliography)
}

}
