% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_top_country.R
\name{plot_top_country}
\alias{plot_top_country}
\title{Used in the pipeline - Plot Top Country}
\usage{
plot_top_country(data)
}
\arguments{
\item{data}{A data frame containing the following columns:
\describe{
\item{Country}{The country of the author.}
\item{Count}{The count of authors from the country.}
}}
}
\value{
A ggplot object representing the bar chart.
}
\description{
This function takes a data frame as input and plots a bar chart of the top countries based on the count of authors.
}
\details{
The \code{plot_top_country} function takes a data frame as input and plots a bar chart of the top countries based on the
count of authors. It filters the data frame to include only countries with a count greater than 10, and then uses
\code{ggplot2} to create the bar chart. The x-axis represents the countries, and the y-axis represents the count of authors.
The title of the plot is "Countries of Institutes of all Authors". The function returns a \code{ggplot} object representing
the bar chart.
}
\examples{
\dontrun{
data <- data.frame(
  Country = c("USA", "Germany", "China", "India"),
  Count = c(20, 15, 12, 10)
)

plot_top_country(data)
}

\dontrun{
data <- data.frame(
  Country = c("USA", "Germany", "China", "India"),
  Count = c(20, 15, 12, 10)
)

plot_top_country(data)
}

}
