% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_toz_country_data.R
\name{plot_top_country_data}
\alias{plot_top_country_data}
\title{Used in the pipeline - Plot Top Country Data}
\usage{
plot_top_country_data(bibliography)
}
\arguments{
\item{bibliography}{A bibliography object containing information about works.
#'}
}
\value{
A data frame with two columns: \verb{Country`` and }Count`.
}
\description{
This function takes a bibliography as input and returns a data frame with the count of works per country.
}
\examples{
\dontrun{
bibliography <- read_bibliography("path/to/bibliography.csv")
plot_top_country_data(bibliography)
}

}
