% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_top_journals.R
\name{plot_top_journals}
\alias{plot_top_journals}
\title{Used in the pipeline - Plot Top Journals}
\usage{
plot_top_journals(data_fn)
}
\arguments{
\item{data}{A data frame containing the journal names and their corresponding citation counts.}
}
\value{
A ggplot object representing the bar chart of the top cited journals.
}
\description{
This function takes a data frame as input and plots a bar chart of the top cited journals.
}
\details{
The \code{plot_top_journals} function takes a data frame as input and creates a bar chart using the \code{ggplot2} package.
The x-axis represents the journal names, ordered by their citation count, and the y-axis represents the count of
citations. The function returns a \code{ggplot} object representing the bar chart.
}
\examples{
\dontrun{
# Create sample data frame
data <- data.frame(
  Journal = c("Journal A", "Journal B", "Journal C"),
  count = c(10, 15, 5)
)

# Plot top journals
plot_top_journals(data)
}

}
