% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_toz_journals_data.R
\name{plot_top_journals_data}
\alias{plot_top_journals_data}
\title{Used in the pipeline - Plot Top Journals Data}
\usage{
plot_top_journals_data(bibliography)
}
\arguments{
\item{bibliography}{A bibliography object.}
}
\value{
A data frame with the top 50 journals and their publication counts.
}
\description{
This function takes a bibliography object and returns a data frame with the
top 50 journals based on the number of publications.
}
\examples{
\dontrun{
bibliography <- read_bibliography("path/to/bibliography.csv")
plot_top_journals_data(bibliography)
}

}
