% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_types_data.R
\name{plot_types_data}
\alias{plot_types_data}
\title{Used in the pipeline - Plot Types Data}
\usage{
plot_types_data(bibliography_fn)
}
\arguments{
\item{bibliography_fn}{A bibliography object containing publication information.}
}
\value{
A data frame with the count of each type of publication from both Zotero and OpenAlex.
}
\description{
This function takes a bibliography object and returns a data frame with the count of each type of publication from both Zotero and OpenAlex.
}
