% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_update_group.R
\name{update_group}
\alias{update_group}
\title{Update group}
\usage{
update_group(group)
}
\arguments{
\item{group}{A named character vector specifying the group names and their
corresponding group IDs.}

\item{overwrite}{Logical value indicating whether to overwrite existing files.
Default is \code{FALSE}.}
}
\description{
This function updates the group by downloading the group data from Zotero.
}
\details{
The \code{update_group} function is used to update the group by downloading the
group data from Zotero. It takes two arguments: \code{group} and \code{overwrite}.

The \code{group} argument is a named character vector that specifies the group names
and their corresponding group IDs. By default, it uses the \code{group()} function
to retrieve the group names and IDs.

The \code{overwrite} argument is a logical value that indicates whether to
overwrite existing files. If set to \code{FALSE}, the function will skip
downloading the group data if the corresponding file already exists. If set to
\code{TRUE}, the function will overwrite existing files.

The function uses the \code{zotero_get_group} function from the \code{IPBES.R}
package to download the group data. It iterates over each group in the \code{group}
vector, downloads the group data, and saves it as a CSV file in the "input"
directory.
The api key for zotero to access private groups can be specified in the environmental
variable \code{ZOTERO_API_KEY}.
}
\examples{
# Update group using default settings
\dontrun{
update_group()

# Update group and overwrite existing files
update_group(overwrite = TRUE)
}

}
